--- close_pipes.c
@@ -1,45 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   close_pipes.c                                      :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: dlevinsc <dlevinsc@student.hive.fi>        +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/08/11 21:37:11 by dlevinsc          #+#    #+#             */
-/*   Updated: 2024/08/25 11:15:23 by dlevinsc         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "minishell.h"
-
-void	close_pipe_fds(t_minishell *shell)
-{
-	int	i;
-
-	i = 0;
-	while (i < shell->pipes_allocated)
-	{
-		if (shell->pipe[i])
-		{
-			close(shell->pipe[i][0]);
-			close(shell->pipe[i][1]);
-		}
-		i++;
-	}
-}
-
-void	close_exec_pipe_fds(t_minishell *shell)
-{
-	int	i;
-
-	i = 0;
-	while (i < shell->pipes_allocated)
-	{
-		if (shell->pipe[i])
-		{
-			close(shell->pipe[i][0]);
-			close(shell->pipe[i][1]);
-		}
-		i++;
-	}
-}
