+++ parsing/parsing_helpers.c
@@ -6,7 +6,7 @@
 /*   By: psitkin <psitkin@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/11 22:15:27 by psitkin           #+#    #+#             */
-/*   Updated: 2024/09/13 15:40:05 by psitkin          ###   ########.fr       */
+/*   Updated: 2024/09/13 15:55:56 by psitkin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -72,12 +72,18 @@ char	*empty_strdup(t_minishell *shell)
 	return (str);
 }
 
-int	ft_strcmp(const char *s1, const char *s2)
+void	child_heredoc(t_minishell *sh, t_cmd_data *cmd, char *file, int i)
 {
-	while (*s1 && *s1 == *s2)
-	{
-		s1++;
-		s2++;
-	}
-	return ((unsigned char)*s1 - (unsigned char)*s2);
+	int		fd;
+
+	signal_toggle(HEREDOC);
+	fd = open(file, O_CREAT | O_RDWR | O_TRUNC, 0644);
+	if (fd == -1)
+		error_p(sh, file, FATAL, 1);
+	write_to_heredoc(sh, cmd->redir[i] + 2, fd);
+	close(fd);
+	all_free(sh);
+	array_free(&sh->env);
+	signal_toggle(DEFAULT);
+	exit(0);
 }
